Applied Eric Just's changes to the Oracle producer. His comments follow:
I have made a couple of more changes which you MIGHT want to incorporate into Oracle Producer so that the translation for CHADO works 'out of the box' .  I am attaching a new diff against the new version.

Change 1:
Change the field 'name' property to the unreserved name, that way whenever field-name is called the proper unreserved name is returned.

Change 2:
Using a CLOB in a unique key is illegal in Oracle.  Convert a column to a varchar(4000) if it is a CLOB and is part of a unique key, warn about this change.

Change 3:
Create the sequence name based on the unreserved name of the table.  Since the table gets created with the unreserved name, this keeps the sequence naming consistently based on the table name.

Change 4:
Similar to change 3 except for naming triggers.

Change 5:
remove line that didn't get removed from last merge.  (want to append name with number only after the string has been truncated to max_length -2)