Added a select section to the parser.  This is to handle function calls that may affect the tables during creation.
Added the setval function.  Unfortunately, nothing is being done with this
because the parser doesn't seem to handle sequences.

Skipping all COPY commands because they are not needed for schema creation.

Added a line each for revoking and granting permissions for a schema.  This
doesn't actually do anything with that information.