Deprecate SQL::Translator::Schema::Graph, undocument as_graph()
According to 10f369208 Producer::Turnkey was retired back in 2004. It seems
that the ::Graph::* family exists solely for Turnkey infrastructure support
and as such can be deprecated. This frees us from a very old dependency
Class::MakeMethods (an early MOP attempt, not updated since 2003).

Fwiw Turnkey itself appears to be also dead:
http://sourceforge.net/mailarchive/forum.php?forum_name=turnkey-users