Default SQLite quoting to off until we are capable of disabling it everywhere
Currently the diffing routines are either "always on" or "always off" with
no ability to pass around a setting (they are all class methods, never
looking at the $translator object). Until we can allow a user to say
quote_identifiers => 0, and not have to change any of the old tests, it
makes little sense to change the default.