don't pass non-JSON specific producer_args to to_json
The producer_args passed by sqlt contains a number of keys which are
not specific to JSON.  These were passed unfiltered to to_json().
JSON (at least as of v2.90) will throw an error if it is passed an
unknown option (it uses the option key as a method name, which leads
to confusing error messages).

It's not straightforward to automatically determine the args supported
by the JSON module, so this simply whitelist the 'pretty', 'indent',
and 'canonical' options.