bugfix: JSON and YAML producers called comments method in wrong context
Both the JSON and YAML producers called the Table and Field comments
method in list context while constructing a hash, assuming they would
return a single value. They don't. In list context the comments
methods return a list of comments.  This led to "odd
number of hash element" errors if the comments method returned an even
number of elements.

In scalar context the comments methods return a string composed of the
concatenation of comments.  This context was explicitly used when
producing procedure comments.

However, since the JSON and YAML formats support arrays, their
producers should output the comments as arrays to avoid loss of
information.

To that end, all comments output are now encoded as arrays.