bugfix: JSON and YAML producers' view_procedure method called comments method in wrong context
The JSON and YAML Producer view_procedure methods called the
procedure's comments() method in scalar context.  However, in scalar
context this method returns a string composed of the
concatenation of comments. Since the JSON and YAML formats support
arrays, their producers should output the comments as arrays to avoid
loss of information.

This is also consistent with the other calls to comment() elsewhere in
the code.