WIP: test quoting of identifiers in diff
In production, we found that the stable CPAN release of the PostgreSQL
producer did not quote identifiers in diffs.  We diagnosed the problem
and wrote a fix, only to see it was fixed in master.  Oh well, we
learned some more about SQLT!

I had adapted my test to run against a few producers, and found that
SQLite didn't work, even in master.  The first cause seemed like the
same bug as I found in PostgreSQL:  add_field did not pass its options
to the generator to inform its quoting.  I have fixed that in the patch,
but it still fails.

The fundamental problem appears to be the weird (to me) `$NO_QUOTES`
behavior.  That variable is set in `->produce`, a method that *does not
seem to be called* in the execution of the test program.  At that point,
I gave up for now, but I wanted to file this issue.