Fix default values returned by Parser::DBI::PostgreSQL
The postgres function pg_get_expr used in reading the default
values for columns always returns a properly quoted DDL expression.
This means that the default values written to $table->add_field
should always be scalar-refs, unless some further step were taken
to parse the DDL for the case of simple strings.

As it was (for the last 17 years?!) the parsed default would be a
string of SQL and if passed back to a producer it would attempt to
quote that SQL as a string value and, in almost all cases, fail.

This was even wrong in the unit tests:

  is( $f3->default_value, "'FOO'::text", 'Default value is "FOO"' );

When written back to the database, 'FOO'::text gets further wrapped
in quotes like '''FOO''::text' resulting in the wrong default value.
for any other data type, it simply fails to create thetable.