Fix data_type and size returned by Parser::DBI::PostgreSQL
The handling of column sizes was also broken for the postgres parser.
It was returning a data_type of 'character varying(50)' with a size
of '54', when it should be returning a type of 'character varying'
and size of '54'.  When passed back to the producer, it generates
syntax errors of "character varying(50)(54)".  For numeric(8,4), it
was ignoring the size entirely.

For reasons undiscovered, the postgres column atttypmod does not
contain useful numbers to represent the size, so I just parse the
size from the data_type and then remove it.  This works for at least
varchar and numeric columns (as shown in the unit test).  It might
not be a universal solution but it is at least a lot more useful
than the previous broken state of things.

While I was at it, I added a check for default-values that are
simple unquoted numbers or simple quoted strings, and un-quote them
to be more useful to the cross-database 'translator' aspect of
SQL::Translator.