Fix bugs in Parser::DBI::PostgreSQL data_type, size, and default_value (#155)
* Fix default values returned by Parser::DBI::PostgreSQL

The postgres function pg_get_expr used in reading the default
values for columns always returns a properly quoted DDL expression.
This means that the default values written to $table->add_field
should always be scalar-refs, unless some further step were taken
to parse the DDL for the case of simple strings.

As it was (for the last 17 years?!) the parsed default would be a
string of SQL and if passed back to a producer it would attempt to
quote that SQL as a string value and, in almost all cases, fail.

This was even wrong in the unit tests:

  is( $f3->default_value, "'FOO'::text", 'Default value is "FOO"' );

When written back to the database, 'FOO'::text gets further wrapped
in quotes like '''FOO''::text' resulting in the wrong default value.
for any other data type, it simply fails to create thetable.

* Fix data_type and size returned by Parser::DBI::PostgreSQL

The handling of column sizes was also broken for the postgres parser.
It was returning a data_type of 'character varying(50)' with a size
of '54', when it should be returning a type of 'character varying'
and size of '54'.  When passed back to the producer, it generates
syntax errors of "character varying(50)(54)".  For numeric(8,4), it
was ignoring the size entirely.

For reasons undiscovered, the postgres column atttypmod does not
contain useful numbers to represent the size, so I just parse the
size from the data_type and then remove it.  This works for at least
varchar and numeric columns (as shown in the unit test).  It might
not be a universal solution but it is at least a lot more useful
than the previous broken state of things.

While I was at it, I added a check for default-values that are
simple unquoted numbers or simple quoted strings, and un-quote them
to be more useful to the cross-database 'translator' aspect of
SQL::Translator.