New ::Parser::DBI::PostgreSQL option deconstruct_enum_types
The ::Producer::PostgreSQL will take a column definition of
data_type 'enum' and convert it to a postgres custom data type,
and then declare the column of that type.  When parsed, the column
no longer shows the values of the enum and the data_type is the
custom name rather than 'enum'.  This new option reaches into the
postgres type to find the enum values, and reverts the data_type
to 'enum', like it was originally.