Handle addition of enum values in ::Producer::PostgreSQL->alter_field
Now, when producing a diff between two enum columns where the list
of values are known, and when the postgres version is greater than
9.001, the producer will generate an ALTER TYPE ... ADD VALUE
statments if the new enum includes values that the old one did not.
This is an easy case to support because it does not involve
re-creating the type.

The harder case of removing an enum value is not supported, yet.
The output includes a SQL comment to that effect, giving the
user a note that they need to perform that step on their own.